Refactors ownerships list into a reusable component
